NF: _fillLrn in Abstract
